requirements.txt
arxiv==1.4.8
openai==0.28.0
python-dotenv==1.0.0
flask==3.0.0
